requirements.txt
google-cloud-bigquery==3.30.0
google-cloud-storage==2.19.0
apache-airflow==2.10.5
pandas==2.2.3
PyYAML==6.0.2
Requests==2.32.3
dbt-core==1.9.2
dbt-bigquery==1.9.1
python-dotenv==1.0.1